+++ app/presentation/api/travel_match_network_system/status.py
@@ -5,6 +5,7 @@ from starlette.websockets import WebSocket
 from app.core.dependencies import DependDriverEventsCase, DependPassengerEventsCase, AuthUserCredentials
 from app.core.types import UUID
 from app.domain.credentials import get_user_credentials_header
+from app.presentation.schemes.status import RideStatus, ListRides, ScheduleStatus
 
 status = APIRouter(prefix="/status", tags=["Status notify"])
 
@@ -29,11 +30,11 @@ async def notify_user_status(uuid: UUID, websocket: WebSocket, events: DependPas
     await events.notify_ws(uuid, websocket, user)
 
 
-@status.get("/events/driver/{uuid}")
-async def events_notify_driver(uuid: UUID, events: DependDriverEventsCase, auth_user: AuthUserCredentials):
+@status.get("/events/driver/{uuid}", response_model=ListRides)
+async def events_notify_driver(uuid: UUID, events: DependDriverEventsCase, auth_user: AuthUserCredentials) -> EventSourceResponse:
     return EventSourceResponse(await events.notify_events(uuid, auth_user))
 
 
-@status.get("/events/passenger/{uuid}")
-async def events_notify_passenger(uuid: UUID, events: DependPassengerEventsCase, auth_user: AuthUserCredentials):
+@status.get("/events/passenger/{uuid}", response_model=ScheduleStatus)
+async def events_notify_passenger(uuid: UUID, events: DependPassengerEventsCase, auth_user: AuthUserCredentials) -> EventSourceResponse:
     return EventSourceResponse(await events.notify_events(uuid, auth_user))
